refactor: extract quake and report mappers into dedicated mapper packages
Move mapping functions out of DataModelExt and into feature-scoped mapper
files (QuakeMapper, ReportMapper) to keep concerns separated by layer.
toCityQuakesListDomain is co-located with CityQuakesEntity.
