+++ app/src/main/java/cl/figonzal/lastquakechile/reports_feature/data/local/entity/CityQuakesEntity.kt
@@ -18,3 +18,5 @@ data class CityQuakesEntity(
         nQuakes = nQuakes
     )
 }
+
+fun List<CityQuakesEntity>.toCityQuakesListDomain() = map { it.toDomain() }
